requirements.txt
anthropic==0.49.0
twilio==9.4.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
requests==2.32.3
flask==3.1.0
python-json-logger==2.0.7

# MFA
pyotp==2.9.0

# PostgreSQL (optional — only needed if DATABASE_URL is set)
psycopg2-binary==2.9.9
